app/consumer_four/requirements.txt
pika==1.2.0
mysql-connector-python==8.0.26
